Command split: Implement the "package" command in a separate class
