+++ lib/yast_module.rb
@@ -6,7 +6,6 @@ class YastModule
   WORK_DIR   = "work"
   RESULT_DIR = "result"
   OBS_DIR = "build_service"
-  OBS_PROJECT = "YaST:Head:ruby"
 
   attr_reader :name,
     :work_dir,
@@ -91,31 +90,6 @@ class YastModule
     exports.map { |e| "#{work_dir}/#{e}/include" }
   end
 
-  def package
-    action "Packaging" do
-      Dir.chdir result_dir do
-        Cheetah.run "make", "-f", "Makefile.cvs" # TODO will not work for cmake based ones
-        Cheetah.run "make"
-        Cheetah.run "make", "package-local", "CHECK_SYNTAX=false"
-      end
-
-      package_name = File.read("#{result_dir}/RPMNAME").chomp
-
-      FileUtils.mkdir_p obs_dir
-      Dir.chdir obs_dir do
-        package_obs_dir = "#{obs_dir}/#{OBS_PROJECT}/#{package_name}"
-
-        create_obs_package_dir!(package_name) unless File.exists?(package_obs_dir)
-
-        Dir["#{result_dir}/package/*"].each do |file|
-          # for last three arguments keep defaults except last that we switch to
-          # overwrite file in destrination
-          FileUtils.copy_entry file, "#{package_obs_dir}/#{File.basename(file)}", false, false, true
-        end
-      end
-    end
-  end
-
   def ybc_module_paths
     # The lazy loading is needed because the dependencies may not be fully
     # initialized when "initialize" is called on this module.
@@ -182,41 +156,6 @@ class YastModule
       :error_ruby  => 0,
       :error_other => 0
     }
-  end
-
-  # cwd must be obs_dir
-  def create_obs_package_dir!(package_name)
-    first = true
-    begin
-      Cheetah.run "osc", "co", OBS_PROJECT, package_name
-    rescue Cheetah::ExecutionFailed
-      # probably it is not created yet, so lets create it
-      if first
-        first = false
-        create_package_in_obs package_name
-        retry
-      end
-
-      raise
-    end
-  end
-
-  def create_package_in_obs(package_name)
-    ENV["EDITOR"] = "ed"
-    # just add description via ed as meta pkg invoke editor
-    ed_command = <<EOS
-      # enable error reporting
-      H
-      # title is required
-      # comma: all lines; s: substitute
-      ,s/<title>/<title>YaST module converted to ruby/
-      # write
-      w
-      # quit
-      q
-EOS
-
-    Cheetah.run "osc", "meta", "pkg", OBS_PROJECT, package_name, "-e", :stdin => StringIO.new(ed_command)
   end
 
   def handle_exception(e, phase, file)
