+++ yk
@@ -28,6 +28,7 @@ require_relative "lib/commands/restructure_command"
 require_relative "lib/commands/ybc_command"
 require_relative "lib/commands/ruby_command"
 require_relative "lib/commands/makefile_command"
+require_relative "lib/commands/package_command"
 
 class CLI < Thor
 
@@ -119,7 +120,7 @@ class CLI < Thor
   def package(*module_names)
     modules = modules_from_names(module_names)
     with_modules modules do |mod|
-      mod.package
+      Commands::PackageCommand.new(@config).apply(mod)
     end
   end
